Lag en egen fil for enum-klassene vi bruker i feilsituasjon da next ikke liker at vi importerer disse fra en dynamisk url i testene
