+++ components/skjema/fullforRegistrering.tsx
@@ -18,7 +18,7 @@ import { useFeatureToggles } from '../../contexts/featuretoggle-context';
 import { useConfig } from '../../contexts/config-context';
 import { Config } from '../../model/config';
 import { hentRegistreringFeiletUrl } from '../../lib/hent-registrering-feilet-url';
-import { OppgaveRegistreringstype } from '../../pages/veiledning/[registreringstype]/[feilsituasjon]';
+import { OppgaveRegistreringstype } from '../../pages/veiledning/feilsituasjonTyper';
 
 const TEKSTER: Tekster<string> = {
     nb: {
