+++ lib/hent-registrering-feilet-url.ts
@@ -1,6 +1,7 @@
 import { ErrorTypes } from '../model/error';
+import { OppgaveRegistreringstype } from '../pages/veiledning/feilsituasjonTyper';
 
-export const hentRegistreringFeiletUrl = (feiltype: ErrorTypes, registreringstype: String) => {
+export const hentRegistreringFeiletUrl = (feiltype: ErrorTypes, registreringstype: OppgaveRegistreringstype) => {
     if (feiltype === ErrorTypes.BRUKER_ER_DOD_UTVANDRET_ELLER_FORSVUNNET) {
         return `/veiledning/${registreringstype}/utvandret/`;
     } else if (feiltype === ErrorTypes.BRUKER_MANGLER_ARBEIDSTILLATELSE) {
