+++ pages/veiledning/feilsituasjonTyper.ts
@@ -1,0 +1,9 @@
+export enum Feiltype {
+    UTVANDRET = 'utvandret',
+    MANGLER_ARBEIDSTILLATELSE = 'mangler-arbeidstillatelse',
+}
+
+export enum OppgaveRegistreringstype {
+    REGISTRERING = 'registrering',
+    REAKTIVERING = 'reaktivering',
+}
